restructured commands folder
